requirements.txt
astroid==2.5.7
atomicwrites==1.4.0
attrs==21.2.0
coverage==5.5
entrypoints==0.3
flake8==3.9.2
importlib-metadata==4.3.1
iniconfig==1.1.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
more-itertools==8.8.0
numpy==1.20.3
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.8.2
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.0
python-dateutil==2.8.1
PyYAML==5.4.1
six==1.16.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
wcwidth==0.2.5
wrapt==1.12.1
wxPython==4.1.1
zipp==3.4.1
